# Optimize integer-->string conversions