@kimikage 2020-07-02
@rfourquet, I appreciate your comment

I added the clamping into the `string()`, and changed the types of the `pad` of internal functions (e.g. `bin()`) from `Integer` to `Int`. Although we can leave them as `Integer` for "compatibility", I think there are no practical advantages.

I don't know the reason, but there is a little slowdown (~1 ns/number) from the original commit, but the unrolling is still effective.