@kimikage 2020-08-08
Since PR #36459 was merged, the `::Int` for `leading_zeros(x)` doesn't seem to be needed anymore.