@rfourquet 2020-10-04
> Perhaps the cause is the type of bit shifting.

@kimikage Thank you for the report, I had no idea this `x << 0x4` was broken in `BitIntegers`.